Move flight/hotel getter outside of main function, prepare it for using different source of data

+++ OTBAssignment/Model/JsonModel/HotelData.cs
@@ -20,7 +20,7 @@ namespace OTBAssignment.Model.JsonModel
         public required IEnumerable<Airport> LocalAirports { get; set; }
 
         [JsonProperty("nights")]
-        public required int Nights { get; set; }
+        public required short Nights { get; set; }
 
     }
 }
